+++ config/postgres_adaptor.py
@@ -43,7 +43,14 @@ class DatabaseConnection:
         return self._session_generator()
 
 
-database_connection = DatabaseConnection(app_config)
+def get_database_connection() -> DatabaseConnection:
+    """
+    Get DB connection.
+
+    Returns:
+        DB connection.
+    """
+    return DatabaseConnection(app_config)
 
 
 async def get_db_session() -> AsyncSession:  # type: ignore
@@ -53,5 +60,6 @@ async def get_db_session() -> AsyncSession:  # type: ignore
     Yields:
         Сессия для работы с БД.
     """
-    async with database_connection.get_session() as session:
+    db_connection = get_database_connection()
+    async with db_connection.get_session() as session:
         yield session