+++ src/database/base.py
@@ -5,19 +5,19 @@ from sqlalchemy.orm import DeclarativeMeta, declarative_base
 Base: DeclarativeMeta = declarative_base()
 
 
-# class UserAlchemyModel(Base):  # type: ignore
-#     """Модель для таблицы урлов."""
-#
-#     __tablename__ = 'urls'
-#
-#     token = Column(
-#         VARCHAR(10),
-#         primary_key=True,
-#         index=True,
-#     )
-#     saved_url = Column(
-#         TEXT,
-#         unique=True,
-#         nullable=False,
-#         index=True,
-#     )
+class UserAlchemyModel(Base):  # type: ignore
+    """Модель для таблицы урлов."""
+
+    __tablename__ = 'urls'
+
+    token = Column(
+        VARCHAR(10),
+        primary_key=True,
+        index=True,
+    )
+    saved_url = Column(
+        TEXT,
+        unique=True,
+        nullable=False,
+        index=True,
+    )